Generation of bpf program
The build binary creates the bpf filter based on the syscalls defined in
struct bpf_call. E.g:
  ./build test.bpf

First, a table with the filtered syscalls is built in ascending order of
syscall number and including the amount of syscalls of that type.
After, the BPF filter with a binary search tree is constructed with:
  1. the nodes for the tree search
  2. the leaves with all the syscall numbers
  3. every syscall arguments if present

Then, the BPF instructions are written in the input file.

Signed-off-by: Alice Frosi <afrosi@redhat.com>